work on proxy code
